demo/final_demo/requirements.txt
numpy==1.20.1
Pillow==8.1.2
torch==1.8.0
torchvision==0.9.0
